% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovizseq.R
\name{biovizseq}
\alias{biovizseq}
\title{BioVizSeq shiny app start function.}
\usage{
biovizseq()
}
\value{
Shinyapp: BioVizSeq shiny app.
}
\description{
BioVizSeq shiny app start function.
}
\examples{
# 1. Library BioVizSeq package
library(BioVizSeq)

}
\author{
Shiqi Zhao
}
