% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdd_plot.R
\name{cdd_to_loc}
\alias{cdd_to_loc}
\title{cdd_to_loc}
\usage{
cdd_to_loc(cdd_file)
}
\arguments{
\item{cdd_file}{CDD file.}
}
\value{
data.frame
}
\description{
Extract the location information of domain from cdd file
}
\examples{
hitdata_path <- system.file("extdata", "hitdata.txt", package = "BioVizSeq")
cdd_file <- readLines(hitdata_path)
domain_loc <- cdd_to_loc(cdd_file)
}
\author{
Shiqi Zhao
}
