% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_plot.R
\name{combi_p}
\alias{combi_p}
\title{combi_p}
\usage{
combi_p(
  tree_path,
  gff_path = NULL,
  meme_path = NULL,
  pfam_path = NULL,
  cdd_path = NULL,
  fa_path = NULL
)
}
\arguments{
\item{tree_path}{The path of tree file (.newick).}

\item{gff_path}{The path of .gff/gtf file.}

\item{meme_path}{The path of .meme/mast file.}

\item{pfam_path}{The path of pfam result file (.tsv).}

\item{cdd_path}{The path of cdd result file (.txt).}

\item{fa_path}{The path of protein file (.fa/fasta).}
}
\value{
list
}
\description{
Get ggplot2 files to facilitate free combination in patchwork
}
\examples{
tree_path <- system.file("extdata", "idpep.nwk", package = "BioVizSeq")
plot_file <- combi_p(tree_path)
}
\author{
Shiqi Zhao
}
