% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdd_plot.R
\name{fastaleng}
\alias{fastaleng}
\title{fastaleng}
\usage{
fastaleng(fasta_file)
}
\arguments{
\item{fasta_file}{The path of protein fasta file.}
}
\value{
data.frame
}
\description{
Statistical sequence length
}
\examples{
fasta_path <- system.file("extdata", "idpep.fa", package = "BioVizSeq") 
fastaleng(fasta_path)

}
\author{
Shiqi Zhao
}
