% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme_plot.R
\name{get_motif_location}
\alias{get_motif_location}
\title{get_motif_location}
\usage{
get_motif_location(motif_file)
}
\arguments{
\item{motif_file}{The motif data of mast or meme file.}
}
\value{
list
}
\description{
Extract the location information of motif from mast or meme file
}
\examples{
meme_path <- system.file("extdata", "meme.xml", package = "BioVizSeq") 
meme_file <- readLines(meme_path)
motif_loc <- get_motif_location(meme_file)

mast_path <- system.file("extdata", "mast.xml", package = "BioVizSeq") 
mast_file <- readLines(mast_path)
motif_loc <- get_motif_location(mast_file)

}
\author{
Shiqi Zhao
}
