% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_plot.R
\name{gff_plot}
\alias{gff_plot}
\title{gff_plot}
\usage{
gff_plot(
  gff_file,
  the_order = NULL,
  shape = "Rect",
  r = 0.3,
  legend_size = 15,
  element_color = NULL
)
}
\arguments{
\item{gff_file}{The path of gff file.}

\item{the_order}{The path of order of mRNA. It is also the mRNA you want to showcase. A List of Gene ID , One ID Per Line.}

\item{shape}{RoundRect or Rect.}

\item{r}{The radius of rounded corners.}

\item{legend_size}{The size of legend.}

\item{element_color}{The color set of element.}
}
\value{
p
}
\description{
Visualization of element in gff or gtf file
}
\examples{
gff_path <- system.file("extdata", "test.gff", package = "BioVizSeq")
gff_plot(gff_path)
}
\author{
Shiqi Zhao
}
