% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvariables.R
\name{globalVariables}
\alias{globalVariables}
\title{Global Variables Declaration}
\description{
Global Variables
}
\details{
This block declares global variables to avoid R CMD check warnings.
}
\keyword{internal}
