% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme_plot.R
\name{meme_seq}
\alias{meme_seq}
\title{meme_seq}
\usage{
meme_seq(meme_file)
}
\arguments{
\item{meme_file}{The path of meme file or mast file.}
}
\value{
data.frame
}
\description{
Get motif sequence from meme file or mast file
}
\examples{
mast_path <- system.file("extdata", "mast.xml", package = "BioVizSeq")
mast_file <- readLines(mast_path)
motifseq<- meme_seq(mast_file)

}
\author{
Shiqi Zhao
}
