% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme_plot.R
\name{meme_to_loc}
\alias{meme_to_loc}
\title{meme_to_loc}
\usage{
meme_to_loc(motif_file)
}
\arguments{
\item{motif_file}{The motif data of mast or meme file.}
}
\value{
list
}
\description{
Extract the location information of motif from mast or meme file
}
\examples{

mast_path <- system.file("extdata", "mast.xml", package = "BioVizSeq") 
mast_file <- readLines(mast_path)
motif_loc <- meme_to_loc(mast_file)

}
\author{
Shiqi Zhao
}
