% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme_plot.R
\name{motif_seq}
\alias{motif_seq}
\title{motif_seq}
\usage{
motif_seq(meme_file)
}
\arguments{
\item{meme_file}{The path of meme file or mast file.}
}
\value{
data.frame
}
\description{
Get motif sequence from meme file or mast file
}
\examples{
meme_path <- system.file("extdata", "meme.xml", package = "BioVizSeq")
meme_file <- readLines(meme_path)
motifseq<- motif_seq(meme_file)

mast_path <- system.file("extdata", "mast.xml", package = "BioVizSeq")
mast_file <- readLines(mast_path)
motifseq<- motif_seq(mast_file)

}
\author{
Shiqi Zhao
}
