% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfam_plot.R
\name{pfam_to_loc}
\alias{pfam_to_loc}
\title{pfam_to_loc}
\usage{
pfam_to_loc(pfam_data)
}
\arguments{
\item{pfam_data}{The result file (.tsv) of pfam (via InterPro).}
}
\value{
list
}
\description{
Extract the location information of domain from pfam result
}
\examples{
pfam_path <- system.file("extdata", "iprscan.tsv", package = "BioVizSeq")
pfam_file <- read.table(pfam_path, sep='\t', header = FALSE)
motif_loc <- pfam_to_loc(pfam_file)
}
\author{
Shiqi Zhao
}
