% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantcare_plot.R
\name{plantcare_classify}
\alias{plantcare_classify}
\title{plantcare_classify}
\usage{
plantcare_classify(plantcare_file)
}
\arguments{
\item{plantcare_file}{The result file (.tab) of Plantcare.}
}
\value{
data.frame
}
\description{
Classify the functions of cis element from Plantcare
}
\examples{
plantcare_path <- system.file("extdata", "plantCARE_output.tab", package = "BioVizSeq") 
plantcare_file <- read.table(plantcare_path, header = FALSE, sep = '\t', quote="")
plantcare_data <- plantcare_classify(plantcare_file)
}
\author{
Shiqi Zhao
}
