% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantcare_plot.R
\name{plantcare_plot3}
\alias{plantcare_plot3}
\title{plantcare_plot3}
\usage{
plantcare_plot3(plantcare_file, the_order = NULL)
}
\arguments{
\item{plantcare_file}{The path of plantcare result file (.tab).}

\item{the_order}{The path of order file. A List of Gene ID , One ID Per Line.}
}
\value{
p
}
\description{
Bar plot2 of the number of cis-element in plantcare result file
}
\examples{
plantcare_path <- system.file("extdata", "plantCARE_output.tab", package = "BioVizSeq") 
plantcare_plot3(plantcare_path)
}
\author{
Shiqi Zhao
}
