% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantcare_plot.R
\name{plantcare_statistic2}
\alias{plantcare_statistic2}
\title{plantcare_statistic2}
\usage{
plantcare_statistic2(plantcare_data)
}
\arguments{
\item{plantcare_data}{The result of plantcare_classify().}
}
\value{
data.frame
}
\description{
Count the number of cis element from Plantcare for Bar chart
}
\examples{
plantcare_path <- system.file("extdata", "plantCARE_output.tab", package = "BioVizSeq") 
plantcare_file <- read.table(plantcare_path, header = FALSE, sep = '\t', quote="")
plantcare_data <- plantcare_classify(plantcare_file)
statistic_data2 <- plantcare_statistic2(plantcare_data)
}
\author{
Shiqi Zhao
}
