% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantcare_plot.R
\name{plantcare_to_loc}
\alias{plantcare_to_loc}
\title{plantcare_to_loc}
\usage{
plantcare_to_loc(plantcare_data)
}
\arguments{
\item{plantcare_data}{The result of plantcare_classify().}
}
\value{
data.frame
}
\description{
Extract the location information of cis-element from Plantcare
}
\examples{
plantcare_path <- system.file("extdata", "plantCARE_output.tab", package = "BioVizSeq") 
plantcare_file <- read.table(plantcare_path, header = FALSE, sep = '\t', quote="")
plantcare_data <- plantcare_classify(plantcare_file)
plantcare_loc <- plantcare_to_loc(plantcare_data)
}
\author{
Shiqi Zhao
}
