% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtParam_calc.R
\name{ProtParam_calc}
\alias{ProtParam_calc}
\title{ProtParam_calc}
\usage{
ProtParam_calc(input_file)
}
\arguments{
\item{input_file}{The path of potein fasta file.}
}
\value{
data.frame
}
\description{
Extract the information of protein sequence
}
\author{
Shiqi Zhao
}
