% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_plot.R
\name{smart_plot}
\alias{smart_plot}
\title{smart_plot}
\usage{
smart_plot(
  fasta_file,
  the_order = NULL,
  domain_select = NULL,
  shape = "RoundRect",
  r = 0.3,
  legend_size = 15,
  domain_color = NULL
)
}
\arguments{
\item{fasta_file}{The path of protein fasta file.}

\item{the_order}{The path of order file. A List of Gene ID , One ID Per Line.}

\item{domain_select}{The domain ID which you want to align with.}

\item{shape}{RoundRect or Rect.}

\item{r}{The radius of rounded corners.}

\item{legend_size}{The size of legend.}

\item{domain_color}{The color set of domain.}
}
\value{
p
}
\description{
Visualization of domain in SMART result file
}
\author{
Shiqi Zhao
}
