% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_plot.R
\name{smart_to_loc}
\alias{smart_to_loc}
\title{smart_to_loc}
\usage{
smart_to_loc(input_file, do_pfam = TRUE)
}
\arguments{
\item{input_file}{The path of potein fasta file.}

\item{do_pfam}{Include the pfam domain or not.}
}
\value{
list
}
\description{
Extract the location information of domain from SMART result
}
\author{
Shiqi Zhao
}
