% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantcare_plot.R
\name{upload_fa_to_plantcare}
\alias{upload_fa_to_plantcare}
\title{upload_fa_to_plantcare}
\usage{
upload_fa_to_plantcare(fasta_file, email)
}
\arguments{
\item{fasta_file}{The path of promoter file.}

\item{email}{e-mail address.}
}
\value{
plantcare_result
}
\description{
Upload the promoter file to Plantcare database
}
\details{
Due to the file size limitation of plantcare on fasta,
upload_fa_to_plantcare() first splits fasta file.
Then uploads the splited fasta files to the plantcare database,
and automatically returns the results to the email provided by the user.

upload_fa_to_plantcare("the path/test.fasta", "your e-mail address")
}
\author{
Shiqi Zhao
}
